src/auth-agent/pyproject.toml
[project]
name = "auth-agent"
version = "1.0.0"
description = "OperatorBoard Auth Service based on agent-framework"
requires-python = ">=3.10"
dependencies = [
    "agent-framework @ file:./agent-framework",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = true
